LayoutTests/http/tests/plugins/third-party-cookie-accept-policy.html

<html>
<head>
<script src="../cookies/resources/cookies-test-pre.js"></script>    
<script>

if (window.layoutTestController) {
    layoutTestController.dumpAsText();
    layoutTestController.waitUntilDone();
    layoutTestController.setAlwaysAcceptCookies(true);
    clearCookies();
    layoutTestController.setAlwaysAcceptCookies(false);
}

</script>
<body>
This tests that plug-ins cannot set cookies in violation of the 3rd party cookie policy.<br>
<iframe src="http://localhost:8000/plugins/resources/third-party-cookie-accept-policy-iframe.html"></iframe>

</body>
</html>
